feat: parser (#17)
# Parsing

Parsing allows you to handle validation failures more gracefully, by
using the `fallback` function:

```ts
type Content = {
  tag: 'text',
  value: string
} | {
  tag: 'unknown'
}

const parseContent = array<Content[]>(
  fallback(
    object({
      tag: literal('text'),
      value: parseString,
    }),
    {
      tag: 'unknown'
    }
  )
)

const res = parseContent([{
  tag: 'text',
  value: 'hello'
}, {
  tag: 'number'
  value: 123
}])
```

Where `res` becomes:

```ts
[
  { tag: 'text', value: 'hello'}, 
  { tag: 'unknown' }
]
```

While a validation function can just evaluate to `false` or `true`, a
parsing result contains more rich information.

This is just the first pull request in a series–it does not provide the
entire solution. Until the parsing is ready for release, it will not be
exported from the package.

Here's the roadmap:

- primitives and literals ("constant" types) 
- unions (sum types)
  - union
  - optional
  - nullable
  - undefinable
  - optionalNullable
-  product types
  - tuples ❌ 
  - object
  - record ❌  
  - partialRecord ❌ 
- recursive types
  -  array
  - nonEmptyArray ❌ 

Lots of tests are missing–many of which are left as todo's.