feat: referential preservation (#35)
## What?

When a value is successfully being parsed, one of two things can happen:

- validation succeeds
- validation fails, but a fallback is returned 

If the validation succeeds, we should always return the original value.
However, this hasn't been the case for product (objects) and recursive
types (arrays) because those were constructed with `.map()`.

The rule is this: if all properties/entries were successfully parsed,
and if all the values in those results were preserved, then also
preserve the value when returning.

These functions needed to be adjusted:

- `union`
- `object`
- `array`
- `fallback`

I.e. pretty much any higher order function. (The `nullable`, `optional`
(etc) are just aliases of `union`, so did not need adjustments.)

Since `fallback()` plays a significant role, I also extracted this
function into its own module, added tests, and fixed a bug (two
fallbacks on top of each other).

## Why?

Preserving the reference is important when working with immutable
libraries, such as React, as this will allow you to utilize memoization.