feat: move failure of missing prop to prop (#80)
## Problem

When using `withDefault` on required properties that are omitted, the
fallback behaviour should take effect, but the current behaviour is that
the failure is yielded at the object level—not at the property—thus the
default is never invoked/

## Solution

Changing the way to handle defaults:

Before: the parser had a special property `[optionalSymbol]`. Whenever
we parsed a property that was missing, we'd check if the parser had this
property.

Now: when the property is missing, we pass a special value
`optionalSymbol` to represent the absence of a property. An optional
parser is now a simple parser as any other. To signify that the value
should be omitted, we return `optionalSymbol`.